Fixed issue in unobserve()
Because the weak_ptr that is created in
~calc_state() is of a derived type of observer_t,
and it is expired, passing it to unobserve will
create an empty weak_ptr<observer_t>, such that
it will not be found. As a result, it will
not be removed from the list of observers,
and it will therefore dangle.

Fortunately, std::owner_less supports arguments
of different typed, as long as they are subobjects
of eachother.

- Fixed by templatizing unobserve() such that
the comparison will be between the weak_ptr<calc_state>
and weak_ptr<observer_t>.