fixed an issue in the constraint of the constructor of calc
for some reason, the constructor was also considered when calling the copy constructor, even though we have a (perfect) copy constructor.
As a result, the derived_from constraint would fail to compile because it would depend on the class itself.

Fixed by first checking whether it is the copy constructor that is being called, by comparing against the class itself.